+++ easy_text/lib/src/core/attributes/builtin/inline/bold_attr.dart
@@ -1,13 +1,13 @@
 import '../../../../../attributes.dart';
 
-class BoldAttribute extends EasyInlineAttribute<bool> {
-  const BoldAttribute([bool value = true]) : super(value: value);
+class BoldAttribute extends EasyInlineAttribute<bool?> {
+  const BoldAttribute([bool? value = true]) : super(value: value);
 
   @override
   String get key => 'strong';
 
   @override
-  BoldAttribute clone(bool value) {
+  BoldAttribute clone(bool? value) {
     return BoldAttribute(value);
   }
 }