+++ easy_text/lib/src/core/attributes/builtin/inline/italic_attr.dart
@@ -1,13 +1,13 @@
 import '../../../../../attributes.dart';
 
-class ItalicAttribute extends EasyInlineAttribute<bool> {
-  const ItalicAttribute([bool value = true]) : super(value: value);
+class ItalicAttribute extends EasyInlineAttribute<bool?> {
+  const ItalicAttribute([bool? value = true]) : super(value: value);
 
   @override
   String get key => 'italic';
 
   @override
-  ItalicAttribute clone(bool value) {
+  ItalicAttribute clone(bool? value) {
     return ItalicAttribute(value);
   }
 }