+++ easy_text/lib/src/core/attributes/builtin/inline/strike_attr.dart
@@ -1,13 +1,13 @@
 import '../../../../../attributes.dart';
 
-class StrikeAttribute extends EasyInlineAttribute<bool> {
-  const StrikeAttribute([bool value = true]) : super(value: value);
+class StrikeAttribute extends EasyInlineAttribute<bool?> {
+  const StrikeAttribute([bool? value = true]) : super(value: value);
 
   @override
   String get key => 'striketrough';
 
   @override
-  StrikeAttribute clone(bool value) {
+  StrikeAttribute clone(bool? value) {
     return StrikeAttribute(value);
   }
 }