+++ easy_text/lib/src/core/attributes/builtin/inline/underline_attr.dart
@@ -1,13 +1,13 @@
 import '../../../../../attributes.dart';
 
-class UnderlineAttribute extends EasyInlineAttribute<bool> {
-  const UnderlineAttribute([bool value = true]) : super(value: value);
+class UnderlineAttribute extends EasyInlineAttribute<bool?> {
+  const UnderlineAttribute([bool? value = true]) : super(value: value);
 
   @override
   String get key => 'underline';
 
   @override
-  UnderlineAttribute clone(bool value) {
+  UnderlineAttribute clone(bool? value) {
     return UnderlineAttribute(value);
   }
 }